Safari-611.1.10/LayoutTests/fast/dom/dataset-name-getter-properties.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body data-webkit-rocks=yes>
<script>
description('Test the JavaScript property descriptor of the name getter.');

shouldBeTrue("Object.getOwnPropertyDescriptor(document.body.dataset, 'webkitRocks').configurable");
shouldBeTrue("Object.getOwnPropertyDescriptor(document.body.dataset, 'webkitRocks').enumerable");
shouldBeTrue("Object.getOwnPropertyDescriptor(document.body.dataset, 'webkitRocks').writable");
shouldBeEqualToString("Object.getOwnPropertyDescriptor(document.body.dataset, 'webkitRocks').value", 'yes');

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
