Safari-611.1.10/LayoutTests/fast/dom/dataset-name-getter-properties-expected.txt

Test the JavaScript property descriptor of the name getter.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Object.getOwnPropertyDescriptor(document.body.dataset, 'webkitRocks').configurable is true
PASS Object.getOwnPropertyDescriptor(document.body.dataset, 'webkitRocks').enumerable is true
PASS Object.getOwnPropertyDescriptor(document.body.dataset, 'webkitRocks').writable is true
PASS Object.getOwnPropertyDescriptor(document.body.dataset, 'webkitRocks').value is "yes"
PASS successfullyParsed is true

TEST COMPLETE
